Crate hydro_deploy

Re-exports

pub use deployment::Deployment;

pub use localhost::LocalhostHost;

pub use gcp::GcpComputeEngineHost;

pub use azure::AzureHost;

pub use aws::AwsEc2Host;

pub use aws::AwsNetwork;

pub use rust_crate::RustCrate;

pub use custom_service::CustomService;

Modules

aws

azure

custom_service

deployment

gcp

localhost

progress

rust_crate

ssh

terraform

util

Structs

ResourceBatch

ResourcePool

ResourceResult

TracingResults

Enums

BaseServerStrategy

ClientStrategy

Like BindType, but includes metadata for determining whether a connection is possible.

HostTargetType

LinuxCompileType

PortNetworkHint

ServerStrategy

Types of connection that a service can receive when configured as the server.

Traits

Host

LaunchedBinary

LaunchedHost

Service

ServiceBuilder

Type Aliases

HostStrategyGetter