Struct hydro_deploy::deployment::Deployment

pub struct Deployment {
    pub hosts: Vec<Weak<dyn Host>>,
    pub services: Vec<Weak<RwLock<dyn Service>>>,
    pub resource_pool: ResourcePool,
    /* private fields */
}

Fields
hosts: Vec<Weak<dyn Host>>services: Vec<Weak<RwLock<dyn Service>>>resource_pool: ResourcePool
Implementations

impl Deployment

pub fn new() -> Self

pub fn Localhost(&self) -> Arc<LocalhostHost>

pub fn CustomService( &mut self, on: Arc<dyn Host>, external_ports: Vec<u16>, ) -> Arc<RwLock<CustomService>>

pub async fn run_until( &mut self, trigger: impl Future<Output = ()>, ) -> Result<()>

Runs deploy(), and start(), waits for the trigger future, then runs stop().

pub async fn start_until( &mut self, trigger: impl Future<Output = ()>, ) -> Result<()>

Runs start(), waits for the trigger future, then runs stop(). This is useful if you need to initiate external network connections between deploy() and start().

pub async fn run_ctrl_c(&mut self) -> Result<()>

Runs deploy(), and start(), waits for CTRL+C, then runs stop().

pub async fn start_ctrl_c(&mut self) -> Result<()>

Runs start(), waits for CTRL+C, then runs stop(). This is useful if you need to initiate external network connections between deploy() and start().

pub async fn deploy(&mut self) -> Result<()>

pub async fn start(&mut self) -> Result<()>

pub async fn stop(&mut self) -> Result<()>

impl Deployment

pub fn add_host<T: Host + 'static, F: FnOnce(usize) -> T>( &mut self, host: F, ) -> Arc<T>

pub fn add_service<T: Service + 'static>( &mut self, service: impl ServiceBuilder<Service = T>, ) -> Arc<RwLock<T>>

impl Deployment

pub fn add_gcp_compute_engine_host( &mut self, project: String, machine_type: String, image: String, region: String, network: Arc<RwLock<GcpNetwork>>, user: Option<String>, startup_script: Option<String>, ) -> Arc<GcpComputeEngineHost>

pub fn add_azure_host( &mut self, project: String, os_type: String, machine_size: String, image: Option<HashMap<String, String>>, region: String, user: Option<String>, ) -> Arc<AzureHost>

impl Deployment

pub fn GcpComputeEngineHost( &mut self, ) -> AddGcpComputeEngineHostDeploymentBuilder<'_>

impl Deployment

pub fn AzureHost(&mut self) -> AddAzureHostDeploymentBuilder<'_>