hydro_deploy::rust_crate::ports

Trait RustCrateSink

pub trait RustCrateSink:
    Any
    + Send
    + Sync {
    // Required methods
    fn instantiate(
        &self,
        client_path: &SourcePath,
    ) -> Result<Box<dyn FnOnce() -> ServerConfig>>;
    fn instantiate_reverse(
        &self,
        server_host: &Arc<dyn Host>,
        server_sink: Arc<dyn RustCrateServer>,
        wrap_client_port: &dyn Fn(ServerConfig) -> ServerConfig,
    ) -> Result<ReverseSinkInstantiator>;
}

Required Methods

fn instantiate( &self, client_path: &SourcePath, ) -> Result<Box<dyn FnOnce() -> ServerConfig>>

Instantiate the sink as the source host connecting to the sink host. Returns a thunk that can be called to perform mutations that instantiate the sink.

fn instantiate_reverse( &self, server_host: &Arc<dyn Host>, server_sink: Arc<dyn RustCrateServer>, wrap_client_port: &dyn Fn(ServerConfig) -> ServerConfig, ) -> Result<ReverseSinkInstantiator>

Instantiate the sink, but as the sink host connecting to the source host. Returns a thunk that can be called to perform mutations that instantiate the sink, taking a mutable reference to this sink.