hydro_deploy

Trait LaunchedBinary

pub trait LaunchedBinary: Send + Sync {
    // Required methods
    fn stdin(&self) -> UnboundedSender<String>;
    fn deploy_stdout(&self) -> Receiver<String>;
    fn stdout(&self) -> UnboundedReceiver<String>;
    fn stderr(&self) -> UnboundedReceiver<String>;
    fn stdout_filter(&self, prefix: String) -> UnboundedReceiver<String>;
    fn stderr_filter(&self, prefix: String) -> UnboundedReceiver<String>;
    fn tracing_results(&self) -> Option<&TracingResults>;
    fn exit_code(&self) -> Option<i32>;
    fn wait<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<i32>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn stop<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn stdin(&self) -> UnboundedSender<String>

fn deploy_stdout(&self) -> Receiver<String>

Provides a oneshot channel to handshake with the binary, with the guarantee that as long as deploy is holding on to a handle, none of the messages will also be broadcast to the user-facing LaunchedBinary::stdout channel.

fn stdout(&self) -> UnboundedReceiver<String>

fn stderr(&self) -> UnboundedReceiver<String>

fn stdout_filter(&self, prefix: String) -> UnboundedReceiver<String>

fn stderr_filter(&self, prefix: String) -> UnboundedReceiver<String>

fn tracing_results(&self) -> Option<&TracingResults>

fn exit_code(&self) -> Option<i32>

fn wait<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<i32>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Wait for the process to stop on its own. Returns the exit code.

fn stop<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

If the process is still running, force stop it. Then run post-run tasks.