Crate hydro_deploy

Re-exports

pub use deployment::Deployment;

pub use localhost::LocalhostHost;

pub use gcp::GcpComputeEngineHost;

pub use azure::AzureHost;

pub use hydroflow_crate::HydroflowCrate;

pub use custom_service::CustomService;

Modules

azure

custom_service

deployment

gcp

hydroflow_crate

localhost

progress

ssh

terraform

util

Structs

ResourceBatch

ResourcePool

ResourceResult

Enums

ClientStrategy

Like BindType, but includes metadata for determining whether a connection is possible.

HostTargetType

ServerStrategy

Types of connections that a host can make to another host.

Traits

Host

LaunchedBinary

LaunchedHost

Service

ServiceBuilder

Type Aliases

HostStrategyGetter