hydro_deploy

Trait Host

pub trait Host:
    Any
    + Send
    + Sync
    + Debug {
    // Required methods
    fn target_type(&self) -> HostTargetType;
    fn request_port_base(&self, bind_type: &BaseServerStrategy);
    fn id(&self) -> usize;
    fn request_custom_binary(&self);
    fn collect_resources(&self, resource_batch: &mut ResourceBatch);
    fn provision(
        &self,
        resource_result: &Arc<ResourceResult>,
    ) -> Arc<dyn LaunchedHost>;
    fn launched(&self) -> Option<Arc<dyn LaunchedHost>>;
    fn strategy_as_server<'a>(
        &'a self,
        connection_from: &dyn Host,
        server_tcp_port_hint: PortNetworkHint,
    ) -> Result<(ClientStrategy<'a>, HostStrategyGetter)>;
    fn can_connect_to(&self, typ: ClientStrategy<'_>) -> bool;

    // Provided method
    fn request_port(&self, bind_type: &ServerStrategy) { ... }
}

Required Methods

fn target_type(&self) -> HostTargetType

fn request_port_base(&self, bind_type: &BaseServerStrategy)

fn id(&self) -> usize

An identifier for this host, which is unique within a deployment.

fn request_custom_binary(&self)

Configures the host to support copying and running a custom binary.

fn collect_resources(&self, resource_batch: &mut ResourceBatch)

Makes requests for physical resources (servers) that this host needs to run.

This should be called before provision is called.

fn provision( &self, resource_result: &Arc<ResourceResult>, ) -> Arc<dyn LaunchedHost>

Connects to the acquired resources and prepares the host to run services.

This should be called after collect_resources is called.

fn launched(&self) -> Option<Arc<dyn LaunchedHost>>

fn strategy_as_server<'a>( &'a self, connection_from: &dyn Host, server_tcp_port_hint: PortNetworkHint, ) -> Result<(ClientStrategy<'a>, HostStrategyGetter)>

Identifies a network type that this host can use for connections if it is the server. The host will be None if the connection is from the same host as the target.

fn can_connect_to(&self, typ: ClientStrategy<'_>) -> bool

Determines whether this host can connect to another host using the given strategy.

Provided Methods

fn request_port(&self, bind_type: &ServerStrategy)